ALL THE EXTENSIONS!!
Add lots of PHP extensions to the SSH image, since it will be used to setup applications (including, but not limited to: Composer, Unit Tests, Functional Tests, Database Migrations, Cache Warming, etc).
Also add many of these extensions to the PHP image so that it is more generic. Will take more time to build, but since it is built locally and not transferred over a network we don't have to worry about image size.